examples/02_voicemail/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
python-multipart==0.0.20
python-dotenv==1.2.1
pyngrok==7.5.1
requests==2.32.5
pydantic==2.9.2
